Print cluster size (;seqs=) as 64-bit
fasta_print_general() took clustersize as int and printed it with %d,
the lone non-64-bit count in a function whose abundance, ordinal and
centroid_size parameters are already uint64_t/int64_t. Widen it to
int64_t and print with PRId64 so the ;seqs= field is not truncated.

(Reaching the limit needs >2^31 sequences in a single cluster; the
caller-side member counts are int and remain the binding limit at that
extreme scale.)

+++ src/fasta.cc
@@ -465,7 +465,7 @@ auto fasta_print_general(std::FILE * output_handle,
                          uint64_t const abundance,
                          int64_t const ordinal,
                          double const expected_error,
-                         int const clustersize,
+                         int64_t const clustersize,
                          int const clusterid,
                          char const * score_name,
                          double const score,
@@ -519,7 +519,7 @@ auto fasta_print_general(std::FILE * output_handle,
 
   if (clustersize > 0)
     {
-      std::fprintf(output_handle, ";seqs=%d", clustersize);
+      std::fprintf(output_handle, ";seqs=%" PRId64, clustersize);
     }
 
   if (clusterid >= 0)
